Rework MockBeanWithPriority as suggested.
