+++ junit4/src/test/java/org/jboss/weld/junit4/bean/AddGloballyEnabledAlternativeTest.java
@@ -1,11 +1,10 @@
 package org.jboss.weld.junit4.bean;
 
-import org.jboss.weld.junit.MockBeanWithPriority;
+import org.jboss.weld.junit.MockBean;
 import org.jboss.weld.junit4.WeldInitiator;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.ArgumentMatchers;
 import org.mockito.Mockito;
 
 import javax.enterprise.inject.spi.Bean;
@@ -14,7 +13,7 @@ import java.util.List;
 import java.util.Set;
 
 /**
- * Tests {@link MockBeanWithPriority} which adds a bean that is always an enabled alternative.
+ * Tests {@link org.jboss.weld.junit.MockBean} adding a bean that is a globally enabled alternative.
  *
  * @author Matej Novotny
  */
@@ -23,20 +22,21 @@ public class AddGloballyEnabledAlternativeTest {
     @Rule
     public WeldInitiator weld = WeldInitiator.from(Bar.class)
             .addBeans(createFooAlternativeBean(), createListBean())
-            .addBeans(MockBeanWithPriority.of(Mockito.mock(MyService.class), MyService.class))
             .build();
 
     public static Bean<?> createFooAlternativeBean() {
-        return MockBeanWithPriority.read(Foo.class)
+        return MockBean.read(Foo.class)
                 .priority(3)
+                .alternative(true)
                 .addQualifier(Meaty.Literal.INSTANCE)
                 .build();
     }
 
     static Bean<?> createListBean() {
-        return MockBeanWithPriority.builder()
+        return MockBean.builder()
                 .types(new TypeLiteral<List<String>>() {
                 }.getType())
+                .globallySelectedAlternative(2)
                 .creating(
                         // Mock object provided by Mockito
                         Mockito.when(Mockito.mock(List.class).get(0)).thenReturn("42").getMock())
@@ -50,17 +50,8 @@ public class AddGloballyEnabledAlternativeTest {
         Assert.assertNotNull(bar.getSomeList());
         Assert.assertEquals("42", bar.getSomeList().get(0));
 
-        // Mock with default settings
-        MyService myService = weld.select(MyService.class).get();
-        myService.doBusiness("Adalbert");
-        Mockito.verify(myService, Mockito.atLeastOnce()).doBusiness(ArgumentMatchers.anyString());
-
         // assert all of these are actually alternatives (enabled)
         Set<Bean<?>> beans = weld.getBeanManager().getBeans(Foo.class, Meaty.Literal.INSTANCE);
-        Assert.assertEquals(1, beans.size());
-        Assert.assertTrue(beans.iterator().next().isAlternative());
-
-        beans = weld.getBeanManager().getBeans(MyService.class);
         Assert.assertEquals(1, beans.size());
         Assert.assertTrue(beans.iterator().next().isAlternative());
 
@@ -68,11 +59,5 @@ public class AddGloballyEnabledAlternativeTest {
         }.getType());
         Assert.assertEquals(1, beans.size());
         Assert.assertTrue(beans.iterator().next().isAlternative());
-    }
-
-    interface MyService {
-
-        void doBusiness(String name);
-
     }
 }
